backend/requirements.txt
aiosmtplib==3.0.2
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
ecdsa==0.19.0
fastapi==0.115.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httptools==0.6.1
idna==3.10
influxdb==5.3.2
influxdb-client==1.46.0
Jinja2==3.1.4
jose==1.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
motor==3.6.0
msgpack==1.1.0
narwhals==1.9.0
packaging==24.1
paho-mqtt==2.1.0
pandas==2.2.3
passlib==1.7.4
pillow==10.4.0
plotly==5.24.1
protobuf==5.28.2
pyarrow==17.0.0
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pymongo==4.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
reactivex==4.0.4
referencing==0.35.1
requests==2.32.3
rich==13.9.1
rpds-py==0.20.0
rsa==4.9
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
streamlit==1.39.0
streamlit-autorefresh==1.0.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.0
watchdog==5.0.3
watchfiles==0.24.0
websockets==13.1

